Split pkg/sync into lifecycle and errorsync
