+++ dispatcher.go
@@ -35,7 +35,7 @@ import (
 	"go.uber.org/yarpc/internal/outboundmiddleware"
 	"go.uber.org/yarpc/internal/pally"
 	"go.uber.org/yarpc/internal/request"
-	intsync "go.uber.org/yarpc/pkg/sync"
+	"go.uber.org/yarpc/pkg/errorsync"
 	"go.uber.org/zap"
 )
 
@@ -301,7 +301,7 @@ func (d *Dispatcher) Start() error {
 
 	abort := func(errs []error) error {
 		// Failed to start so stop everything that was started.
-		wait := intsync.ErrorWaiter{}
+		wait := errorsync.ErrorWaiter{}
 		for _, s := range allStarted {
 			wait.Submit(s.Stop)
 		}
@@ -319,7 +319,7 @@ func (d *Dispatcher) Start() error {
 	d.log.Debug("Set router for inbounds.")
 
 	// Start Transports
-	wait := intsync.ErrorWaiter{}
+	wait := errorsync.ErrorWaiter{}
 	d.log.Debug("Starting transports.")
 	for _, t := range d.transports {
 		wait.Submit(start(t))
@@ -330,7 +330,7 @@ func (d *Dispatcher) Start() error {
 	d.log.Debug("Started transports.")
 
 	// Start Outbounds
-	wait = intsync.ErrorWaiter{}
+	wait = errorsync.ErrorWaiter{}
 	d.log.Debug("Starting outbounds.")
 	for _, o := range d.outbounds {
 		wait.Submit(start(o.Unary))
@@ -342,7 +342,7 @@ func (d *Dispatcher) Start() error {
 	d.log.Debug("Started outbounds.")
 
 	// Start Inbounds
-	wait = intsync.ErrorWaiter{}
+	wait = errorsync.ErrorWaiter{}
 	d.log.Debug("Starting inbounds.")
 	for _, i := range d.inbounds {
 		wait.Submit(start(i))
@@ -381,7 +381,7 @@ func (d *Dispatcher) Stop() error {
 
 	// Stop Inbounds
 	d.log.Debug("Stopping inbounds.")
-	wait := intsync.ErrorWaiter{}
+	wait := errorsync.ErrorWaiter{}
 	for _, i := range d.inbounds {
 		wait.Submit(i.Stop)
 	}
@@ -392,7 +392,7 @@ func (d *Dispatcher) Stop() error {
 
 	// Stop Outbounds
 	d.log.Debug("Stopping outbounds.")
-	wait = intsync.ErrorWaiter{}
+	wait = errorsync.ErrorWaiter{}
 	for _, o := range d.outbounds {
 		if o.Unary != nil {
 			wait.Submit(o.Unary.Stop)
@@ -408,7 +408,7 @@ func (d *Dispatcher) Stop() error {
 
 	// Stop Transports
 	d.log.Debug("Stopping transports.")
-	wait = intsync.ErrorWaiter{}
+	wait = errorsync.ErrorWaiter{}
 	for _, t := range d.transports {
 		wait.Submit(t.Stop)
 	}
