+++ peer/single.go
@@ -27,12 +27,12 @@ import (
 	"go.uber.org/yarpc/api/peer"
 	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/internal/introspection"
-	intsync "go.uber.org/yarpc/pkg/sync"
+	"go.uber.org/yarpc/pkg/lifecycle"
 )
 
 // Single implements the Chooser interface for a single peer
 type Single struct {
-	once          intsync.LifecycleOnce
+	once          lifecycle.LifecycleOnce
 	t             peer.Transport
 	pid           peer.Identifier
 	p             peer.Peer
@@ -43,7 +43,7 @@ type Single struct {
 // NewSingle creates a static Chooser with a single Peer
 func NewSingle(pid peer.Identifier, transport peer.Transport) *Single {
 	s := &Single{
-		once: intsync.Once(),
+		once: lifecycle.Once(),
 		pid:  pid,
 		t:    transport,
 	}
