+++ transport/http/transport.go
@@ -32,7 +32,7 @@ import (
 	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/internal/backoff"
 	"go.uber.org/yarpc/peer/hostport"
-	intsync "go.uber.org/yarpc/pkg/sync"
+	"go.uber.org/yarpc/pkg/lifecycle"
 )
 
 type transportOptions struct {
@@ -134,7 +134,7 @@ func NewTransport(opts ...TransportOption) *Transport {
 
 func (o *transportOptions) newTransport() *Transport {
 	return &Transport{
-		once:                intsync.Once(),
+		once:                lifecycle.Once(),
 		client:              o.buildClient(o),
 		connTimeout:         o.connTimeout,
 		connBackoffStrategy: o.connBackoffStrategy,
@@ -164,7 +164,7 @@ func buildHTTPClient(options *transportOptions) *http.Client {
 // services and pooling the resources needed therein.
 type Transport struct {
 	lock sync.Mutex
-	once intsync.LifecycleOnce
+	once lifecycle.LifecycleOnce
 
 	client *http.Client
 	peers  map[string]*httpPeer
